Adding null check for devices

+++ Atspm/ConfigApi/Controllers/LocationController.cs
@@ -149,7 +149,7 @@ namespace Utah.Udot.Atspm.ConfigApi.Controllers
                 {
                     //get the version previous to the one being deleted
                     var previousVersion = versions.Where(w => w.Start < location.Start).OrderByDescending(o => o.Start).FirstOrDefault();
-                    if (previousVersion != null)
+                    if (previousVersion != null && location.Devices != null)
                     {
                         //assign the devices of the deleted version to the location id of the previous version
                         foreach (var device in location.Devices)
@@ -160,13 +160,16 @@ namespace Utah.Udot.Atspm.ConfigApi.Controllers
                 }
                 else
                 {
-                    foreach (var device in location.Devices)
+                    if (location.Devices != null)
                     {
-                        _deviceRepository.Remove(device);
+                        foreach (var device in location.Devices)
+                        {
+                            _deviceRepository.Remove(device);
+                        }
                     }
                 }
                 await _repository.SetLocationToDeleted(key);
-            }                                
+            }
             catch (ArgumentException e)
             {
                 return NotFound(e.Message);
